Added tavern
